Server/requirements.txt
# Core dependencies
paho-mqtt==1.6.1
flask==2.3.3
numpy==1.24.3
pandas==2.0.3

# Machine Learning
tensorflow==2.13.0
scikit-learn==1.3.0
joblib==1.3.2

# Data & API
requests==2.31.0
sqlalchemy==2.0.20
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
loguru==0.7.0